+++ axelor-account/src/main/java/com/axelor/apps/account/service/SubrogationReleaseServiceImpl.java
@@ -217,6 +217,12 @@ public class SubrogationReleaseServiceImpl implements SubrogationReleaseService 
         continue;
       }
 
+      boolean isRefund = false;
+      if(invoice.getOperationTypeSelect() == InvoiceRepository.OPERATION_TYPE_CLIENT_REFUND) 
+      {
+    	  isRefund=true;
+      }
+      
       LocalDate date = subrogationRelease.getAccountingDate();
       Move move =
           moveService
@@ -239,7 +245,7 @@ public class SubrogationReleaseServiceImpl implements SubrogationReleaseService 
                   invoice.getPartner(),
                   factorDebitAccount,
                   invoice.getCompanyInTaxTotalRemaining(),
-                  true,
+                  !isRefund,
                   date,
                   null,
                   1,
@@ -254,7 +260,7 @@ public class SubrogationReleaseServiceImpl implements SubrogationReleaseService 
                   invoice.getPartner(),
                   factorCreditAccount,
                   invoice.getCompanyInTaxTotalRemaining(),
-                  false,
+                  isRefund,
                   date,
                   null,
                   2,